Perofromance-improvement: Clear choosableElements
Clearing this huge array of PsiElements as soon as possible to free memory.

NOTE: This is not the best solution but a temporary fix. Having this many PsiElements in memory is bad for performance - but doing it step by step would mean that you cannot undo it in one step either.
Idea: Try using stubs instead of PsiElements. But I did not find a good example of that yet.